requirements.txt
azure-functions
flask
redis
pymongo[srv]
python-dotenv
